gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
bcrypt>=4.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
# watermarking
numpy>=1.26.0
scipy>=1.12.0
Pillow>=10.0.0
# test dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiosqlite>=0.20.0
